requirements.txt
flask 
firebase-admin 
opencv-python 
pandas 
pyzbar 
pillow